requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
